+++ Distribution/LuaBridge/LuaBridge.h
@@ -2878,6 +2878,163 @@ ReturnType doCall(T* obj, MemFn& fn, TypeListValues<Params>& tvl)
 
 //=================================================================================================
 /**
+ * @brief Generic function traits.
+ */
+template <bool IsMember, bool IsConst, class R, class... Args>
+struct function_traits_base
+{
+    using result_type = R;
+
+    using argument_types = std::tuple<Args...>;
+
+    static constexpr auto arity = sizeof...(Args);
+
+    static constexpr auto is_member = IsMember;
+
+    static constexpr auto is_const = IsConst;
+};
+
+template <class...>
+struct function_traits_impl;
+
+template <class R, class... Args>
+struct function_traits_impl<R(Args...)> : function_traits_base<false, false, R, Args...>
+{
+};
+
+template <class R, class... Args>
+struct function_traits_impl<R (*)(Args...)> : function_traits_base<false, false, R, Args...>
+{
+};
+
+template <class C, class R, class... Args>
+struct function_traits_impl<R (C::*)(Args...)> : function_traits_base<true, false, R, Args...>
+{
+};
+
+template <class C, class R, class... Args>
+struct function_traits_impl<R (C::*)(Args...) const> : function_traits_base<true, true, R, Args...>
+{
+};
+
+template <class R, class... Args>
+struct function_traits_impl<R(Args...) noexcept> : function_traits_base<false, false, R, Args...>
+{
+};
+
+template <class R, class... Args>
+struct function_traits_impl<R (*)(Args...) noexcept> : function_traits_base<false, false, R, Args...>
+{
+};
+
+template <class C, class R, class... Args>
+struct function_traits_impl<R (C::*)(Args...) noexcept> : function_traits_base<true, false, R, Args...>
+{
+};
+
+template <class C, class R, class... Args>
+struct function_traits_impl<R (C::*)(Args...) const noexcept> : function_traits_base<true, true, R, Args...>
+{
+};
+
+template <class F>
+struct functor_traits_impl : function_traits_impl<decltype(&F::operator())>
+{
+};
+
+/**
+ * @brief Traits class for callable objects (e.g. function pointers, lambdas)
+ *
+ * @tparam F Callable object.
+ */
+template <class F>
+struct function_traits : std::conditional<std::is_class<F>::value,
+                                          detail::functor_traits_impl<F>,
+                                          detail::function_traits_impl<F>>::type
+{
+};
+
+/**
+ * @brief Deduces the return type of a callble object.
+ *
+ * @tparam F Callable object.
+ */
+template <class F>
+using function_result_t = typename function_traits<F>::result_type;
+
+/**
+ * @brief Deduces the argument type of a callble object.
+ *
+ * @tparam I Argument index.
+ * @tparam F Callable object.
+ */
+template <std::size_t I, class F>
+using function_argument_t = std::tuple_element_t<I, typename function_traits<F>::argument_types>;
+
+/**
+ * @brief An integral constant expression that gives the number of arguments accepted by the callable object.
+ *
+ * @tparam F Callable object.
+ */
+template <class F>
+static constexpr std::size_t function_arity_v = function_traits<F>::arity;
+
+/**
+ * @brief An boolean constant expression that checks if the callable object is a member function.
+ *
+ * @tparam F Callable object.
+ */
+template <class F>
+static constexpr bool function_is_member_v = function_traits<F>::is_member;
+
+/**
+ * @brief An boolean constant expression that checks if the callable object is const.
+ *
+ * @tparam F Callable object.
+ */
+template <class F>
+static constexpr bool function_is_const_v = function_traits<F>::is_const;
+
+/**
+ * @brief Detect if we are a `std::function`.
+ */
+template <class F, class...>
+struct is_std_function : std::false_type
+{
+};
+
+template <class ReturnType, class... Args>
+struct is_std_function<std::function<ReturnType(Args...)>> : std::true_type
+{
+};
+
+template <class Signature>
+struct is_std_function<std::function<Signature>> : std::true_type
+{
+};
+
+template <class F>
+static constexpr bool is_std_function_v = is_std_function<F>::value;
+
+/**
+ * @brief Reconstruct a function signature from return type and args.
+ */
+template <class, class>
+struct to_function_type
+{
+};
+
+template <class ReturnType, typename... Ts>
+struct to_function_type<ReturnType, std::tuple<Ts...>>
+{
+    using type = std::function<ReturnType(Ts...)>;
+};
+
+template <class ReturnType, typename... Ts>
+using to_function_type_t = typename to_function_type<ReturnType, Ts...>::type;
+
+//=================================================================================================
+/**
  * @brief Traits for function pointers.
  *
  * There are three types of functions: global, non-const member, and const member. These templates determine the type of function, which
@@ -5695,6 +5852,22 @@ class Namespace : public detail::Registrar
 
         //--------------------------------------------------------------------------
         /**
+            Add or replace a namespace function by convertible to std::function (capturing lambdas).
+        */
+        template <class Function, typename = std::enable_if_t<! detail::is_std_function_v<Function>>>
+        Class<T>& addFunction(char const* name, Function function)
+        {
+            using FnTraits = detail::function_traits<Function>;
+            
+            using FnType = detail::to_function_type_t<
+                typename FnTraits::result_type,
+                typename FnTraits::argument_types>;
+            
+            return addFunction(name, FnType(function));
+        }
+
+        //--------------------------------------------------------------------------
+        /**
             Add or replace a member function by std::function.
         */
         template <class ReturnType, class... Params>
@@ -5715,7 +5888,7 @@ class Namespace : public detail::Registrar
 
             return *this;
         }
-
+        
         //--------------------------------------------------------------------------
         /**
             Add or replace a const member function by std::function.
@@ -6171,6 +6344,22 @@ public:
         }
 
         return *this;
+    }
+
+    //----------------------------------------------------------------------------
+    /**
+        Add or replace a namespace function by convertible to std::function (capturing lambdas).
+    */
+    template <class Function, typename = std::enable_if_t<! detail::is_std_function_v<Function>>>
+    Namespace& addFunction(char const* name, Function function)
+    {
+        using FnTraits = detail::function_traits<Function>;
+        
+        using FnType = detail::to_function_type_t<
+            typename FnTraits::result_type,
+            typename FnTraits::argument_types>;
+        
+        return addFunction(name, FnType(function));
     }
 
     //----------------------------------------------------------------------------